move ClientSession to a separate file
